Switch to UTC for all timestamps
Events are already being stored in UTC, but unless the browser and
server are in the same timezone, timestamps get munged.  Using
`picker.setLocalDate` is especially problematic.  With this commit,
everything is entered and displayed in UTC.

The timezone *can* be made configurable, and it would probably make
sense to allow users to enter times in their local timezone, but
that's too much work for right now.